Refactor tests and improve error handling for codecs
