+++ .env.example
@@ -2,7 +2,3 @@
 XMTP_ENV="dev" # xmtp environment (dev,production,local,multinode)
 LOGGING_LEVEL="error" # rust library logs
 LOG_LEVEL="debug" # JS logs level
-
-# TESTS
-BATCH_SIZE=5-10-15
-DELIVERY_AMOUNT=10
